Trying OperatorPolicy support
just a very quick proof of concept to support
OperatorPolicy.

Not intendt to create a real implementation. Just
patching and adding ifs here and there to just
make TALM to not brake when this Kind appears.

Actually, TALM does not support different Policies
Kinds. Only ConfigurationPolicies, so, I added
many quicks conditions.

Only for demoing if it makes sense to support it.